Move setting wiring to FB3 dialect
