+++ src/NHibernate/Dialect/FirebirdDialect.cs
@@ -8,7 +8,6 @@ using NHibernate.Dialect.Schema;
 using NHibernate.Engine;
 using NHibernate.SqlCommand;
 using NHibernate.Type;
-using NHibernate.Util;
 using Environment = NHibernate.Cfg.Environment;
 
 namespace NHibernate.Dialect
@@ -42,29 +41,6 @@ namespace NHibernate.Dialect
 		public override string AddColumnString
 		{
 			get { return "add"; }
-		}
-
-		/// <summary>
-		/// Set by <see cref="Environment.FirebirdUseNativeBoolean" />.
-		/// </summary>
-		public bool UseNativeBoolean { get; private set; }
-
-		/// <inheritdoc />
-		public override void Configure(IDictionary<string, string> settings)
-		{
-			base.Configure(settings);
-
-			UseNativeBoolean = PropertiesHelper.GetBoolean(Environment.FirebirdUseNativeBoolean, settings);
-			if (UseNativeBoolean)
-				RegisterColumnType(DbType.Boolean, "BOOLEAN");
-		}
-
-		/// <inheritdoc />
-		public override string ToBooleanValueString(bool value)
-		{
-			return UseNativeBoolean
-				? value ? "true" : "false"
-				: base.ToBooleanValueString(value);
 		}
 
 		public override string GetSelectSequenceNextValString(string sequenceName)
