+++ src/NHibernate/PublicAPI/PublicAPI.Unshipped.txt
@@ -1,9 +1,10 @@
 *REMOVED*override NHibernate.Id.SequenceIdentityGenerator.SequenceIdentityDelegate.PrepareIdentifierGeneratingInsert() -> NHibernate.Id.Insert.IdentifierGeneratingInsert
 NHibernate.Dialect.Firebird3Dialect
 NHibernate.Dialect.Firebird3Dialect.Firebird3Dialect() -> void
-NHibernate.Dialect.FirebirdDialect.UseNativeBoolean.get -> bool
+NHibernate.Dialect.Firebird3Dialect.UseNativeBoolean.get -> bool
 const NHibernate.Cfg.Environment.FirebirdUseNativeBoolean = "firebird.use_native_boolean" -> string
 override NHibernate.Dialect.Firebird3Dialect.AppendIdentitySelectToInsert(NHibernate.SqlCommand.SqlString insertString, string identifierColumnName) -> NHibernate.SqlCommand.SqlString
+override NHibernate.Dialect.Firebird3Dialect.Configure(System.Collections.Generic.IDictionary<string, string> settings) -> void
 override NHibernate.Dialect.Firebird3Dialect.GetCreateSequenceString(string sequenceName) -> string
 override NHibernate.Dialect.Firebird3Dialect.GetDropSequenceString(string sequenceName) -> string
 override NHibernate.Dialect.Firebird3Dialect.GetLimitString(NHibernate.SqlCommand.SqlString queryString, NHibernate.SqlCommand.SqlString offset, NHibernate.SqlCommand.SqlString limit) -> NHibernate.SqlCommand.SqlString
@@ -15,8 +16,7 @@ override NHibernate.Dialect.Firebird3Dialect.NoColumnsInsertString.get -> string
 override NHibernate.Dialect.Firebird3Dialect.SupportsIdentityColumns.get -> bool
 override NHibernate.Dialect.Firebird3Dialect.SupportsInsertSelectIdentity.get -> bool
 override NHibernate.Dialect.Firebird3Dialect.SupportsPooledSequences.get -> bool
-override NHibernate.Dialect.FirebirdDialect.Configure(System.Collections.Generic.IDictionary<string, string> settings) -> void
-override NHibernate.Dialect.FirebirdDialect.ToBooleanValueString(bool value) -> string
+override NHibernate.Dialect.Firebird3Dialect.ToBooleanValueString(bool value) -> string
 override NHibernate.Dialect.MySQL8Dialect.ForUpdateNowaitString.get -> string
 override NHibernate.Dialect.MySQL8Dialect.GetForUpdateNowaitString(string aliases) -> string
 override NHibernate.Dialect.MySQL8Dialect.GetForUpdateString(string aliases) -> string
